requirements.txt
# Core
numpy
pandas
scikit-learn
xgboost
tensorflow
prophet
cmdstanpy


# Visualization
matplotlib
seaborn
streamlit
streamlit-autocomplete
streamlit-toggle
streamlit-toggle-switch
plotly


# Web scraping & API
requests
beautifulsoup4
lxml

# Trends & Social Media
tweepy
pytrends

# Env management
python-dotenv

# Time handling
holidays
pycountry

# Jupyter (for notebooks folder)
notebook

# Testing
pytest
